pyproject.toml
[tool.poetry]
name = "hh-api-db"
version = "0.1.0"
description = ""
authors = ["MaxZ <maxzonoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
types-psycopg2 = "^2.9.21.20241019"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.1"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0.20241016"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
